Snow Validator Class' inherits from R6Class

This class is used to validate snow data.

Value

a list with lenient and strict validation

Public fields

STRICT_PATTERN

Regular expression patterns for strict validation'

LENIENT_PATTERN

Regular expression patterns for lenient validation

PATTERNS

Regular expression patterns for validating individual fields

Methods

Public methods

SnowValidator$new()

SnowValidator$validate()

SnowValidator$validateField()

SnowValidator$clone()

Method new()

Regular expression patterns for strict validation

Regular expression patterns for strict validation

Regular expression patterns for validating individual fields

Initialize the SnowValidator class

Usage

SnowValidator$new()

Returns

None

Method validate()

Validate a line of snow data

Usage

SnowValidator$validate(line)

Arguments

line

A string representing a line of snow data

Returns

A list with lenient and strict validation results

Method validateField()

Validate individual fields in a line of snow data

Usage

SnowValidator$validateField(line)

Arguments

line

A string representing a line of snow data

Returns

A list indicating the validity of each field

Method clone()

The objects of this class are cloneable with this method.

Usage

SnowValidator$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

snow_validator <- SnowValidator$new()
snow_validator$validate("NS-20230101-001 60.123456N 15.123456E 100.0 os")
#> $strict
#> [1] FALSE
#> 
#> $lenient
#> [1] TRUE
#> 
#> $fields
#> $fields$prefix
#> [1] TRUE
#> 
#> $fields$datetime
#> [1] TRUE
#> 
#> $fields$latitude
#> [1] TRUE
#> 
#> $fields$longitude
#> [1] TRUE
#> 
#> $fields$elevation
#> [1] TRUE
#> 
#> $fields$obs_code
#> [1] TRUE
#> 
#>